+++ packages/destination-actions/src/destinations/hubspot/index.ts
@@ -6,7 +6,6 @@ import upsertContact from './upsertContact'
 import upsertCompany from './upsertCompany'
 import upsertCustomObjectRecord from './upsertCustomObjectRecord'
 import upsertObject from './upsertObject'
-import upsertObject2 from './upsertObject2'
 import customEvent from './customEvent'
 import { HUBSPOT_BASE_URL } from './properties'
 interface RefreshTokenResponse {
@@ -61,8 +60,7 @@ const destination: DestinationDefinition<Settings> = {
     upsertCompany,
     upsertCustomObjectRecord,
     upsertObject,
-    customEvent,
-    upsertObject2
+    customEvent
   }
 }
