+++ packages/destination-actions/src/destinations/hubspot/upsertObject/index.ts
@@ -1,8 +1,10 @@
-import type { ActionDefinition, RequestClient } from '@segment/actions-core'
+import { ActionDefinition, RequestClient, IntegrationError } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import { HubspotClient, AssociationSyncMode, SyncMode } from './hubspot-api'
 import { commonFields } from './common-fields'
+import { SchemaMatch } from './hubspot-api'
+
 import {
   dynamicReadAssociationLabels,
   dynamicReadIdFields,
@@ -12,9 +14,9 @@ import {
 } from './dynamic-fields'
 
 const action: ActionDefinition<Settings, Payload> = {
-  title: 'Upsert Object',
+  title: 'Custom Object',
   description:
-    'Upsert a record of any Object type to HubSpot and optionally assocate it with another record of any Object type.',
+    'Add, create or update records of any Object type to HubSpot, and optionally assocate that record with other records of any Object type.',
   syncMode: {
     description: 'Specify how Segment should update Records in Hubspot',
     label: 'Sync Mode',
@@ -30,11 +32,11 @@ const action: ActionDefinition<Settings, Payload> = {
   },
   dynamicFields: {
     object_details: {
-      from_object_type: async (request) => {
+      object_type: async (request) => {
         return await dynamicReadObjectTypes(request)
       },
-      from_id_field_name: async (request, { payload }) => {
-        const fromObjectType = payload?.object_details?.from_object_type
+      id_field_name: async (request, { payload }) => {
+        const fromObjectType = payload?.object_details?.object_type
 
         if (!fromObjectType) {
           throw new Error("Select from 'From Object Type' first")
@@ -42,8 +44,8 @@ const action: ActionDefinition<Settings, Payload> = {
 
         return await dynamicReadIdFields(request, fromObjectType)
       },
-      from_property_group: async (request, { payload }) => {
-        const fromObjectType = payload?.object_details?.from_object_type
+      property_group: async (request, { payload }) => {
+        const fromObjectType = payload?.object_details?.object_type
 
         if (!fromObjectType) {
           throw new Error("Select from 'From Object Type' first")
@@ -54,7 +56,7 @@ const action: ActionDefinition<Settings, Payload> = {
     },
     properties: {
       __keys__: async (request, { payload }) => {
-        const fromObjectType = payload?.object_details?.from_object_type
+        const fromObjectType = payload?.object_details?.object_type
 
         if (!fromObjectType) {
           throw new Error("Select from 'From Object Type' first")
@@ -63,9 +65,9 @@ const action: ActionDefinition<Settings, Payload> = {
         return await dynamicReadProperties(request, fromObjectType, false)
       }
     },
-    sensitiveProperties: {
+    sensitive_properties: {
       __keys__: async (request, { payload }) => {
-        const fromObjectType = payload?.object_details?.from_object_type
+        const fromObjectType = payload?.object_details?.object_type
 
         if (!fromObjectType) {
           throw new Error("Select from 'From Object Type' first")
@@ -75,7 +77,7 @@ const action: ActionDefinition<Settings, Payload> = {
       }
     },
     associations: {
-      to_object_type: async (request) => {
+      object_type: async (request) => {
         return await dynamicReadObjectTypes(request)
       },
       association_label: async (request, { dynamicFieldContext, payload }) => {
@@ -85,8 +87,8 @@ const action: ActionDefinition<Settings, Payload> = {
           throw new Error('Selected array index is missing')
         }
 
-        const fromObjectType = payload?.object_details?.from_object_type
-        const toObjectType = payload?.associations?.[selectedIndex]?.to_object_type
+        const fromObjectType = payload?.object_details?.object_type
+        const toObjectType = payload?.associations?.[selectedIndex]?.object_type
 
         if (!fromObjectType) {
           throw new Error("Select from 'From Object Type' first")
@@ -98,14 +100,14 @@ const action: ActionDefinition<Settings, Payload> = {
 
         return await dynamicReadAssociationLabels(request, fromObjectType, toObjectType)
       },
-      to_id_field_name: async (request, { dynamicFieldContext, payload }) => {
+      id_field_name: async (request, { dynamicFieldContext, payload }) => {
         const selectedIndex = dynamicFieldContext?.selectedArrayIndex
 
         if (selectedIndex === undefined) {
           throw new Error('Selected array index is missing')
         }
 
-        const toObjectType = payload?.associations?.[selectedIndex]?.to_object_type
+        const toObjectType = payload?.associations?.[selectedIndex]?.object_type
 
         if (!toObjectType) {
           throw new Error("Select from 'To Object Type' first")
@@ -125,41 +127,55 @@ const action: ActionDefinition<Settings, Payload> = {
 
 const send = async (request: RequestClient, payloads: Payload[], syncMode: SyncMode) => {
   const {
-    object_details: {
-      from_object_type: fromObjectType,
-      from_id_field_name: fromIdFieldName,
-      from_property_group: fromPropertyGroup
-    },
+    object_details: { object_type: objectType, property_group: propertyGroup },
     association_sync_mode: assocationSyncMode
   } = payloads[0]
 
   const client = new HubspotClient(
     request,
-    fromObjectType,
-    fromIdFieldName,
+    objectType,
     syncMode,
     assocationSyncMode as AssociationSyncMode,
-    fromPropertyGroup
+    propertyGroup
   )
 
-  const cleanedPayloads = client.cleanProperties(payloads)
+  const cleanedPayloads = client.cleanProps(payloads)
+  const schema = client.schema(cleanedPayloads)
+  const schemaDiffCache = await client.schemaDiffCache(schema)
 
-  if (fromPropertyGroup) {
-    const { uniqueProperties, uniqueSensitiveProperties } = client.uniquePayloadsProperties(cleanedPayloads)
-    const { properties: propertiesFromHSchema, sensitiveProperties: sensitivePropertiesFromHSchema } =
-      await client.propertiesFromHSchema(uniqueProperties.length > 0, uniqueSensitiveProperties.length > 0)
-    const propertiesToCreate = client.propertiesToCreateInHSSchema(uniqueProperties, propertiesFromHSchema)
-    const sensitivePropertiesToCreate = client.propertiesToCreateInHSSchema(
-      uniqueSensitiveProperties,
-      sensitivePropertiesFromHSchema
-    )
-    await client.ensurePropertiesInHSSchema(propertiesToCreate, sensitivePropertiesToCreate)
+  switch (schemaDiffCache.match) {
+    case SchemaMatch.FullMatch: {
+      const fromRecordPayloads = await client.sendFromRecords(cleanedPayloads)
+      const associationPayloads = client.createAssociationPayloads(fromRecordPayloads)
+      const associatedRecords = await client.sendAssociatedRecords(associationPayloads)
+      await client.sendAssociations(associatedRecords)
+      return
+    }
+    case SchemaMatch.PropertiesMissing:
+    case SchemaMatch.NoMatch: {
+      const schemaDiffHubspot = await client.schemaDiffHubspot(schema)
+
+      switch (schemaDiffHubspot.match) {
+        case SchemaMatch.FullMatch: {
+          await client.saveSchemaToCache(schema)
+          break
+        }
+        case SchemaMatch.PropertiesMissing: {
+          await client.createProperties(schemaDiffHubspot)
+          await client.saveSchemaToCache(schema)
+          break
+        }
+        case SchemaMatch.NoMatch: {
+          throw new IntegrationError('Object Type missing', 'Object Type missing', 400)
+        }
+      }
+
+      const fromRecordPayloads = await client.sendFromRecords(cleanedPayloads)
+      const associationPayloads = client.createAssociationPayloads(fromRecordPayloads)
+      const associatedRecords = await client.sendAssociatedRecords(associationPayloads)
+      await client.sendAssociations(associatedRecords)
+    }
   }
-  
-  const fromRecordsOnHS = await client.ensureFromRecordsOnHubspot(cleanedPayloads)
-  const associations = client.getAssociationsFromPayloads(fromRecordsOnHS)
-  const toRecordsOnHS = await client.ensureToRecordsOnHubspot(associations)
-  await client.ensureAssociations(toRecordsOnHS)
 }
 
 export default action
