--- packages/destination-actions/src/destinations/hubspot/upsertObject2/__tests__/index.test.ts
@@ -1,5 +1,0 @@
-describe('HubSpot.customObject', () => {
-  it('empty test', async () => {
-    expect(true).toBe(true)
-  })
-})
