--- packages/destination-actions/src/destinations/hubspot/upsertObject2/constants.ts
@@ -1,19 +1,0 @@
-export const INSERT_TYPES = [
-  { label: 'Create', value: 'create' },
-  { label: 'Update', value: 'update' },
-  { label: 'Upsert', value: 'upsert' }
-]
-
-export const SUPPORTED_HUBSPOT_OBJECT_TYPES = [
-  { label: 'Contact', value: 'contact' },
-  { label: 'Company', value: 'company' },
-  { label: 'Deal', value: 'deal' },
-  { label: 'Ticket', value: 'ticket' },
-  { label: 'Calls', value: 'calls' },
-  { label: 'Emails', value: 'email' },
-  { label: 'Meetings', value: 'meetings' },
-  { label: 'Postal Mail', value: 'postal_mail' },
-  { label: 'Tasks', value: 'tasks' }
-]
-
-export const MAX_HUBSPOT_BATCH_SIZE = 100
